+++ apps/web-nomic/src/screens/account_details/utils.tsx
@@ -1,73 +1,50 @@
-import chainConfig from '@/chainConfig';
 import {
-  AccountBalancesDocument,
-  // AccountWithdrawalAddressDocument,
-  AccountDelegationBalanceDocument,
-} from '@/graphql/general/account_details_documents';
-import axios from 'axios';
+  useAccountBalancesQuery,
+  // useAccountWithdrawalAddressQuery,
+  useAccountDelegationBalanceQuery,
+} from '@/graphql/types/general_types';
 
-const { endpoints } = chainConfig();
-const urlEndpoints = [
-  process.env.NEXT_PUBLIC_GRAPHQL_URL,
-  endpoints.graphql,
-  'http://localhost:3000/v1/graphql',
-];
-
-export const fetchAvailableBalances = async (address: string) => {
+export const useAvailableBalances = (address?: string) => {
   const defaultReturnValue = {
     accountBalances: {
       coins: [],
     },
   };
-  try {
-    const url = urlEndpoints.find((u) => u) ?? '';
-    const { data } = await axios.post(url, {
-      variables: {
-        address,
-      },
-      query: AccountBalancesDocument,
-    });
-    return data?.data ?? defaultReturnValue;
-  } catch (error) {
-    return defaultReturnValue;
-  }
+  const { data } = useAccountBalancesQuery({
+    variables: {
+      address: address ?? '',
+    },
+    skip: !address,
+  });
+  return data ?? defaultReturnValue;
 };
 
-// export const fetchAccountWithdrawalAddress = async (address: string) => {
+// export const useAccountWithdrawalAddress = (address?: string) => {
 //   const defaultReturnValue = {
 //     withdrawalAddress: {
 //       address,
 //     },
 //   };
-//   try {
-//     const { data } = await axios.post(getUrl(), {
-//       variables: {
-//         address,
-//       },
-//       query: AccountWithdrawalAddressDocument,
-//     });
-//     return data?.data ?? defaultReturnValue;
-//   } catch (error) {
-//     return defaultReturnValue;
-//   }
+//   const { data } = useAccountWithdrawalAddressQuery({
+//     variables: {
+//       address: address ?? '',
+//     },
+//     skip: !address,
+//   });
+//   return data ?? defaultReturnValue;
 // };
 
-export const fetchDelegationBalance = async (address: string) => {
+export const useDelegationBalance = (address?: string) => {
   const defaultReturnValue = {
     delegationBalance: {
       coins: [],
     },
   };
-  try {
-    const url = urlEndpoints.find((u) => u) ?? '';
-    const { data } = await axios.post(url, {
-      variables: {
-        address,
-      },
-      query: AccountDelegationBalanceDocument,
-    });
-    return data?.data ?? defaultReturnValue;
-  } catch (error) {
-    return defaultReturnValue;
-  }
+  const { data } = useAccountDelegationBalanceQuery({
+    variables: {
+      address: address ?? '',
+    },
+    skip: !address,
+  });
+  return data ?? defaultReturnValue;
 };