+++ packages/ui/src/screens/home/components/blocks/types.ts
@@ -7,6 +7,7 @@ export interface BlockType {
 }
 
 export interface BlocksState {
+  loading: boolean;
   items: BlockType[];
 }