+++ packages/ui/src/screens/home/components/transactions/hooks.ts
@@ -9,6 +9,7 @@ import { useState } from 'react';
 
 export const useTransactions = () => {
   const [state, setState] = useState<TransactionsState>({
+    loading: true,
     items: [],
   });
 
@@ -17,8 +18,12 @@ export const useTransactions = () => {
   // ================================
   useTransactionsListenerSubscription({
     onData: (data) => {
-      setState({
-        items: data.data.data ? formatTransactions(data.data.data) : [],
+      setState((prevState) => {
+        const newState = {
+          loading: false,
+          items: data.data.data ? formatTransactions(data.data.data) : [],
+        };
+        return R.equals(prevState, newState) ? prevState : newState;
       });
     },
   });