+++ packages/ui/src/screens/profile_details/hooks.test.tsx
@@ -19,84 +19,51 @@ jest.mock('next/router', () => ({
 }));
 
 jest.mock('@/hooks', () => ({
-  useDesmosProfile: (options: Parameters<typeof useDesmosProfile>[0]) => ({
-    fetchDesmosProfile: jest.fn((dtag) => {
-      let results;
-      if (dtag === '@happieSa') {
-        results = {
-          address: 'desmos18tug2x5uwkgnh7qgadezvdntpwgjc88c98zuck',
-          bio: 'hungry all the time',
-          dtag: 'HappieSa',
-          nickname: 'theHappySamoyed',
-          chainLinks: [],
-          applicationLinks: [],
-          creationTime: '2021-10-06T00:10:45.761731',
-          coverPic:
-            'https://ipfs.desmos.network/ipfs/Qmf48cpgi2zNiH24Vo1xtVsePUJx9665gtiRduVCvV5fFg',
-          profilePic:
-            'https://ipfs.desmos.network/ipfs/QmTvkdGrtBHHihjVajqqA2HAoHangeKR1oYbQWzasnPi7B',
-        };
-      }
-      if (dtag === '@forbole') {
-        results = {
-          address: 'desmos1pm6pmpsdw8kd5g6jneyq8rl3qw6tukcp7g57w3',
-          bio: 'Forbole [ˈfɔːbəl] is a well-established blockchain validator and developer since 2017.',
-          dtag: 'forbole',
-          nickname: 'Forbole',
-          chainLinks: [],
-          applicationLinks: [],
-          creationTime: '2021-10-06T00:10:45.761731',
-          coverPic:
-            'https://ipfs.desmos.network/ipfs/Qmf48cpgi2zNiH24Vo1xtVsePUJx9665gtiRduVCvV5fFg',
-          profilePic:
-            'https://ipfs.desmos.network/ipfs/QmTvkdGrtBHHihjVajqqA2HAoHangeKR1oYbQWzasnPi7B',
-        };
-      }
+  ...jest.requireActual('@/hooks'),
+  useDesmosProfile: (options: Parameters<typeof useDesmosProfile>[0]) => {
+    if (options.addresses?.[0] === '@happieSa') {
+      return {
+        data: [
+          {
+            bio: 'hungry all the time',
+            dtag: 'HappieSa',
+            nickname: 'theHappySamoyed',
+            chainLinks: [],
+            applicationLinks: [],
+            creationTime: '2021-10-06T00:10:45.761731',
+            coverPic:
+              'https://ipfs.desmos.network/ipfs/Qmf48cpgi2zNiH24Vo1xtVsePUJx9665gtiRduVCvV5fFg',
+            profilePic:
+              'https://ipfs.desmos.network/ipfs/QmTvkdGrtBHHihjVajqqA2HAoHangeKR1oYbQWzasnPi7B',
+            connections: [{ identifier: `${prefix.account}test` }],
+          },
+        ],
+        loading: false,
+      };
+    }
+    if (options.addresses?.[0] === '@forbole') {
+      return {
+        data: [
+          {
+            bio: 'Forbole [ˈfɔːbəl] is a well-established blockchain validator and developer since 2017.',
+            dtag: 'forbole',
+            nickname: 'Forbole',
+            chainLinks: [],
+            applicationLinks: [],
+            creationTime: '2021-10-06T00:10:45.761731',
+            coverPic:
+              'https://ipfs.desmos.network/ipfs/Qmf48cpgi2zNiH24Vo1xtVsePUJx9665gtiRduVCvV5fFg',
+            profilePic:
+              'https://ipfs.desmos.network/ipfs/QmTvkdGrtBHHihjVajqqA2HAoHangeKR1oYbQWzasnPi7B',
+            connections: [{ identifier: `${prefix.account}test` }],
+          },
+        ],
+        loading: false,
+      };
+    }
 
-      return options.onComplete({
-        profile: results ? [results] : [],
-      });
-    }),
-    formatDesmosProfile: jest.fn((data) => {
-      let results;
-      if (data?.profile?.[0]?.dtag === 'HappieSa') {
-        results = {
-          dtag: 'HappieSa',
-          nickname: 'theHappySamoyed',
-          imageUrl:
-            'https://ipfs.desmos.network/ipfs/Qmf48cpgi2zNiH24Vo1xtVsePUJx9665gtiRduVCvV5fFg',
-          coverUrl:
-            'https://ipfs.desmos.network/ipfs/QmTvkdGrtBHHihjVajqqA2HAoHangeKR1oYbQWzasnPi7B',
-          bio: 'hungry all the time',
-          connections: [
-            {
-              network: 'native',
-              identifier: `${prefix.account}1kmw9et4e99ascgdw0mmkt63mggjuu0xuqjx30w`,
-            },
-          ],
-        };
-      }
-
-      if (data?.profile?.[0]?.dtag === 'forbole') {
-        results = {
-          dtag: 'forbole',
-          nickname: 'Forbole',
-          imageUrl:
-            'https://ipfs.desmos.network/ipfs/Qmf48cpgi2zNiH24Vo1xtVsePUJx9665gtiRduVCvV5fFg',
-          coverUrl:
-            'https://ipfs.desmos.network/ipfs/QmTvkdGrtBHHihjVajqqA2HAoHangeKR1oYbQWzasnPi7B',
-          bio: 'Forbole [ˈfɔːbəl] is a well-established blockchain validator and developer since 2017.',
-          connections: [
-            {
-              network: 'native',
-              identifier: `${prefix.account}1pm6pmpsdw8kd5g6jneyq8rl3qw6tukcp7g57w3`,
-            },
-          ],
-        };
-      }
-      return results;
-    }),
-  }),
+    return { data: [], loading: false };
+  },
 }));
 
 describe('hook: useProfileDetails', () => {